task-78318 update services and config, add tests

+++ cinema/src/main/java/ru/job4j/persistence/Store.java
@@ -17,5 +17,5 @@ public interface Store {
 
     Collection<Ticket> getAllTickets();
 
-    Optional<Long> createOrder(Collection<Integer> ticketIds, Account customer);
+    Optional<Long> createOrder(Collection<Long> ticketIds, Account customer);
 }
